refactor: remove SimplePageShell and consolidate into CanvasPage
- Deleted SimplePageShell component and its associated tests.
- Updated index exports to include CanvasPage instead of SimplePageShell.
- Modified page-shells test to validate CanvasPage export.
- Introduced CanvasPage component to unify layout for scroll-parallax pages, combining features of SimplePageShell and DocsPageShell.
- Implemented IntersectionObserver for active section tracking in both title-bar and sections modes.
- Updated rendering logic to support both title-bar and sections modes within CanvasPage.
